+++ lib/spack/spack/cmd/test.py
@@ -96,7 +96,7 @@ def test(parser, args, unknown_args):
         pytest.main(['-h'])
         return
 
-    # pytest.ini lives in the root of the spack repository.
+    # pytest.ini lives in lib/spack/spack/test
     with working_dir(spack.paths.test_path):
         # --list and --long-list print the test output better.
         if args.list or args.long_list:
